requirements.txt
chromadb
openai
python-dotenv
langchain_community
langchain_text_splitters
pypdf
flask
streamlit
pycryptodome==3.15.0